+++ lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:crypto/screens/wallet/wallet_page.dart';
+import 'package:crypto/wallet/wallet_page.dart';
 import 'package:device_preview/device_preview.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
