+++ lib/wallet/wallet_page.dart
@@ -1,0 +1,15 @@
+import 'package:crypto/wallet/wallet_screen.dart';
+import 'package:flutter/material.dart';
+import '../../shared/widgets/bottom_navigation_bar.dart';
+
+class WalletPage extends StatelessWidget {
+  const WalletPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: WalletScreen(),
+      bottomNavigationBar: BottomNavBar(),
+    );
+  }
+}
